unsubscription-are-free/pyproject.toml
[project]
name = "unsubscription-are-free"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pwntools>=4.14.1",
]
